297. Serialize and Deserialize Binary Tree
This was pretty easy since I only needed to port and tweak the code
I already wrote for TreeUtils. The idea is simple - use a level
order traversal to serialize, but include the null child values for
any non-null node.

The deserialzation code will look at the length of the "parts" and
use that as a guide as it basically re-creates the level order
traversal that producted the serialized string.

If we were using something like Strings instead of Integers use
the "Unit Separator" ASCII character = 0x1F (31) instead.